use drawLines instead of drawLine

+++ app/src/main/java/com/froloapp/telegramchart/widget/linechartview/LineChartView.java
@@ -157,7 +157,7 @@ public class LineChartView extends AbsLineChartView {
                     stampInfoPaint.setColor(chart.getColor());
                     stampInfoPaint.setAlpha((int) (alpha * 255));
                     int index = adapter.getTimestampIndex(xAxis);
-                    long value = chart.getValueAt(index);
+                    float value = chart.getValueAt(index);
                     float y = getYCoor(value);
                     canvas.drawCircle(x, y, stampInfoBigDotRadius, stampInfoPaint);
                     stampInfoPaint.setColor(stampInfoSmallDotColor);
